+++ azur-engine/src/BulletSpawnerComponent.h
@@ -14,8 +14,8 @@ public:
 private:
 	PositionComponent* position;
 	int frame;
-	int bullet_angle = 0;
-	int full_pattern_angle_offset = 0;
+	float bullet_angle = 0;
+	float full_pattern_angle_offset = 0;
 };
